requirements.txt
streamlit==1.32.0
pandas==2.2.0
mysql-connector-python==8.3.0
plotly==5.19.0
requests==2.31.0
python-dotenv==1.0.1
openai==1.12.0
anthropic==0.18.1